fix(indexing): avoid Kilo embedding dimension override
Kilo-hosted embedding models use Cloud catalog dimensions for vector-store setup, but the embedding request itself should not forward a dimensions override through the gateway. Some upstream embedding APIs reject that parameter even when it matches the native dimension.

Also move the Kilo embedding catalog retry helper into a JSX-free module so unit tests can import it in CI without requiring the Solid JSX runtime, and tighten the retry-cap assertion.
