+++ packages/kilo-indexing/src/indexing/embedders/kilo.ts
@@ -18,7 +18,6 @@ export class KiloEmbedder implements IEmbedder {
     baseUrl?: string
     organizationId?: string
     modelId?: string
-    dimensions?: number
   }) {
     if (!input.apiKey) throw new Error("Kilo API key is required for embedding.")
 
@@ -34,7 +33,7 @@ export class KiloEmbedder implements IEmbedder {
       input.apiKey,
       this.model,
       MAX_ITEM_TOKENS,
-      { headers, dimensions: input.dimensions },
+      { headers },
     )
   }
